Fix a bug when creating a new gallery page
I was using a method to convert from a relative to absolute path that depended on the file actually existing. However, as it was returning the path for the file that was actually being created, that file didn't exist yet. I've updated the method to use a different method that only requires the directory the file is in to exist.