Fix bug in finding image size
The path being used to find an image's size was in the output folder when it should have been in the content folder. This was causing failures any time the image file hadn't yet been written to content. Now image_tag correctly locates the image file in content when determining image size.